Adding another example macromodel file
This new file (meets1.mae) will reproduce an error when
trying to compare the -mt and -jt torsions. This is due to an
error we know happens when grabbing torsional data from the mmo file.
Ideally we should have a check through the code to say some torsions
are not being picked up, and which ones. Both the ref and the calc
data list should also be trimmed and aligned based off of the label.